move remote model to baseline/model.py
